PFM-1099: Clean up project as per suggestion

+++ backend/muster-roll-service/src/main/java/org/egov/web/models/AttendanceEntry.java
@@ -29,7 +29,7 @@ import lombok.Builder;
 @Builder
 public class AttendanceEntry   {
         @JsonProperty("time")
-        private BigDecimal time = null;
+        private Long time = null;
 
         @JsonProperty("attendance")
         private BigDecimal attendance = null;
